Conflicts:
	lsstsal/scripts/code/templates/SALDDS.cpp.template
	lsstsal/scripts/gensalgetput.tcl

+++ lsstsal/scripts/runtime-generator.tcl
@@ -1,11 +1,7 @@
 #!/usr/bin/env tclsh
 
-puts stdout "Removing old idl-templates"
-exec rm -fr idl-templates
-
 source $env(SAL_DIR)/add_system_dictionary.tcl
 set EVERYTHING [lsort $SYSDIC(systems)]
-
 foreach subsys $EVERYTHING {
    catch {
       set x [glob $subsys*.xml]
@@ -15,9 +11,13 @@ foreach subsys $EVERYTHING {
    }
 }
 
-puts stdout  "Validating interfaces"
-foreach subsys $EVERYTHING {
-  if { [info exists DO($subsys)] } {
+if { $argv == "" || [lsearch $argv validate] > -1} {
+  puts stdout "Removing old idl-templates"
+  exec rm -fr idl-templates
+
+  puts stdout  "Validating interfaces"
+  foreach subsys $EVERYTHING {
+   if { [info exists DO($subsys)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys validate ] } bad
@@ -26,51 +26,63 @@ foreach subsys $EVERYTHING {
    set result ""
    catch {exec salgenerator $subsys html}
    puts stdout "$result $bad"
+   }
   }
 }
 
-puts stdout  "Generating C++"
-foreach subsys $EVERYTHING {
+if { $argv == "" || [lsearch $argv cpp] > -1 } {
+ puts stdout  "Generating C++"
+ foreach subsys $EVERYTHING {
   if { [info exists DO($subsys)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys sal cpp ] } bad
    puts stdout "$result $bad"
   }
+ }
 }
 
-puts stdout  "Generating Python"
-foreach subsys $EVERYTHING {
+if { $argv == "" || [lsearch $argv python] > -1 } {
+ puts stdout  "Generating Python"
+ foreach subsys $EVERYTHING {
   if { [info exists DO($subsys)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys sal python ] } bad
    puts stdout "$result $bad"
   }
+ }
 }
 
-puts stdout  "Generating LabVIEW"
-foreach subsys $EVERYTHING {
+
+if { $argv == "" || [lsearch $argv labview] > -1 } {
+ puts stdout  "Generating LabVIEW"
+ foreach subsys $EVERYTHING {
   if { [info exists DO($subsys)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys labview ] } bad
    puts stdout "$result $bad"
   }
+ }
 }
 
-puts stdout  "Updating libraries"
-foreach subsys $EVERYTHING {
+
+if { $argv == "" || [lsearch $argv lib] > -1 } {
+ puts stdout  "Updating libraries"
+ foreach subsys $EVERYTHING {
   if { [info exists DO($subsys)] } {
    set bad ""
    set result ""
    catch { set results [exec salgenerator $subsys lib ] } bad
    puts stdout "$result $bad"
   }
+ }
 }
 
-puts stdout  "Generating Java"
-foreach subsys $EVERYTHING {
+if { $argv == "" || [lsearch $argv java] > -1 } {
+ puts stdout  "Generating Java"
+ foreach subsys $EVERYTHING {
   if { [info exists DO($subsys)] } {
    exec rm -fr [set subsys]/java
    set bad ""
@@ -83,13 +95,59 @@ foreach subsys $EVERYTHING {
    puts stdout "$result $bad"
 
   }
+ }
 }
 
-source $env(SAL_DIR)/gengenericefd.tcl
-set bad ""
-set result ""
-catch { set results [updateefdschema] } bad
-puts stdout "$result $bad"
+
+if { $argv == "" || [lsearch $argv efd] > -1 } {
+ puts stdout  "Updating EFD SQL writers"
+ source $env(SAL_DIR)/gengenericefd_array.tcl
+ set bad ""
+ set result ""
+ catch { set results [updateefdschema] } bad
+ puts stdout "$result $bad"
+ puts stdout  "Updating EFD InfluxDB writers"
+ source $env(SAL_DIR)/geninfluxefd-multi.tcl
+ set bad ""
+ set result ""
+ catch { set results [updateinfluxschema] } bad
+ puts stdout "$result $bad"
+ puts stdout  "Updating EFD kafka writers"
+ source $env(SAL_DIR)/genkafkaefd.tcl
+ set bad ""
+ set result ""
+ catch { set results [updatekafkaschema] } bad
+ puts stdout "$result $bad"
+}
+
+
+
+cd $env(SAL_WORK_DIR)
+
+if { $argv == "" || [lsearch $argv rpm] > -1 } {
+ puts stdout  "Updating RPMs"
+ foreach subsys $EVERYTHING {
+  if { [info exists DO($subsys)] } {
+   set bad ""
+   set result ""
+   catch { set results [exec salgenerator $subsys rpm] } bad
+   puts stdout "$result $bad"
+  }
+ }
+ source $env(SAL_DIR)/gensalrpms.tcl
+ set bad ""
+ set result ""
+ catch { set results [generatemetarpm] } bad
+ puts stdout "$result $bad"
+ set bad ""
+ set result ""
+ catch { set results [generateATmetarpm] } bad
+ puts stdout "$result $bad"
+ set bad ""
+ set result ""
+ catch { set results [generateEFDrpm] } bad
+ puts stdout "$result $bad"
+}
